# docs/openapi.yaml
openapi: 3.0.3
info:
  title: AuthKit API
  version: 0.1.0-rc.1
  description: Static integration contract. Runtime documentation endpoints are intentionally disabled.
servers:
  - url: https://auth.example.com
security:
  - bearerAuth: []
paths:
...
  /api/v1/admin/users:
    get:
      parameters:
        - {name: search, in: query, required: false, schema: {type: string, maxLength: 255}, description: Case-insensitive email or display-name search}
        - {name: limit, in: query, required: false, schema: {type: integer, minimum: 1, maximum: 200, default: 100}}
        - {name: cursor, in: query, required: false, schema: {type: string, maxLength: 1024}, description: Signed short-lived cursor bound to search and limit}
      responses: {'200': {$ref: '#/components/responses/Success'}, '400': {$ref: '#/components/responses/Error'}, '401': {$ref: '#/components/responses/Error'}, '403': {$ref: '#/components/responses/Error'}}
  /api/v1/admin/users/{userId}:
    get:
      parameters: [{$ref: '#/components/parameters/UserId'}]
      description: Safe account and authenticator status; no password hashes, provider subjects, credential IDs, public keys, or MFA secrets are returned.
      responses: {'200': {$ref: '#/components/responses/Success'}, '401': {$ref: '#/components/responses/Error'}, '403': {$ref: '#/components/responses/Error'}, '404': {$ref: '#/components/responses/Error'}}
  /api/v1/admin/users/{userId}/sessions/revoke:
    post:
      parameters: [{$ref: '#/components/parameters/UserId'}]
      description: Revokes every AuthKit session for the target after recent local administrator step-up.
      requestBody: {$ref: '#/components/requestBodies/MfaOptionalVerificationRequired'}
      responses: {'200': {$ref: '#/components/responses/Success'}, '400': {$ref: '#/components/responses/Error'}, '401': {$ref: '#/components/responses/Error'}, '403': {$ref: '#/components/responses/Error'}}
  /api/v1/admin/users/{userId}/role:
    patch:
      parameters: [{$ref: '#/components/parameters/UserId'}]
      requestBody: {$ref: '#/components/requestBodies/AdminUpdateRole'}
      responses: {'200': {$ref: '#/components/responses/Success'}}
  /api/v1/admin/users/{userId}/suspension:
    post:
      parameters: [{$ref: '#/components/parameters/UserId'}]
      requestBody: {$ref: '#/components/requestBodies/AdminAccountState'}
      responses: {'200': {$ref: '#/components/responses/Success'}}
    delete:
      parameters: [{$ref: '#/components/parameters/UserId'}]
      requestBody: {$ref: '#/components/requestBodies/AdminAccountState'}
      responses: {'200': {$ref: '#/components/responses/Success'}}
  /api/v1/admin/users/{userId}/deletion:
    delete:
      parameters: [{$ref: '#/components/parameters/UserId'}]
      requestBody: {$ref: '#/components/requestBodies/AdminAccountState'}
      responses: {'200': {$ref: '#/components/responses/Success'}}
  /api/v1/admin/oauth-clients:
    get: *authGet
    post:
      requestBody: {$ref: '#/components/requestBodies/AdminOAuthClientCreate'}
      responses: {'200': {$ref: '#/components/responses/Success'}, '400': {$ref: '#/components/responses/Error'}}
  /api/v1/admin/oauth-clients/{clientId}:
    patch:
      parameters: [{$ref: '#/components/parameters/ClientId'}]
      requestBody: {$ref: '#/components/requestBodies/AdminOAuthClientUpdate'}
      responses: {'200': {$ref: '#/components/responses/Success'}}
    delete:
      parameters: [{$ref: '#/components/parameters/ClientId'}]
      requestBody: {$ref: '#/components/requestBodies/MfaOptionalVerification'}
      responses: {'200': {$ref: '#/components/responses/Success'}}
  /api/v1/admin/social-providers:
    get: *authGet
    post:
      requestBody:
        required: true
        content: {application/json: {schema: {$ref: '#/components/schemas/AdminSocialProviderCreateRequest'}}}
      responses: {'200': {$ref: '#/components/responses/Success'}, '400': {$ref: '#/components/responses/Error'}}
  /api/v1/admin/social-providers/{providerId}:
    patch:
      parameters: [{$ref: '#/components/parameters/ProviderId'}]
      requestBody: {$ref: '#/components/requestBodies/AdminSocialProviderUpdate'}
      responses: {'200': {$ref: '#/components/responses/Success'}}
    delete:
      parameters: [{$ref: '#/components/parameters/ProviderId'}]
      requestBody: {$ref: '#/components/requestBodies/MfaOptionalVerificationRequired'}
      responses: {'200': {$ref: '#/components/responses/Success'}}
  /api/v1/admin/security-events:
    get:
      parameters:
        - {name: userId, in: query, required: false, schema: {type: string, format: uuid}}
        - {name: limit, in: query, required: false, schema: {type: integer, minimum: 1, maximum: 200, default: 100}}
        - {name: cursor, in: query, required: false, schema: {type: string, maxLength: 1024}}
      description: Privacy-minimized security-event view with masked network and identity fields.
      responses: {'200': {$ref: '#/components/responses/Success'}, '400': {$ref: '#/components/responses/Error'}, '401': {$ref: '#/components/responses/Error'}, '403': {$ref: '#/components/responses/Error'}}
  /api/v1/admin/operations:
    get:
      description: Non-secret operational posture and provider-acceptance outbox counts.
      responses: {'200': {$ref: '#/components/responses/Success'}, '401': {$ref: '#/components/responses/Error'}, '403': {$ref: '#/components/responses/Error'}}
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: First-party AuthKit access token with `token_use=first_party_access`, configured API audience, non-revoked key, and active server-side session. OAuth access tokens and ID tokens are rejected.
    oauthAccessBearer:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: OAuth access token with `token_use=oauth_access`, OAuth client audience, `client_id`, and server-side scope checks. ID tokens and first-party access tokens are rejected.
    refreshCookie:
      type: apiKey
      in: cookie
      name: Refresh-Token
      description: HttpOnly refresh cookie scoped to AuthKit refresh/logout routes.
    csrfHeader:
      type: apiKey
      in: header
      name: X-XSRF-TOKEN
      description: CSRF header that must echo the readable CSRF cookie on refresh and logout.
    oauthClientBasic:
      type: http
      scheme: basic
      description: OAuth client authentication for token, introspection, and revocation endpoints. Public clients may use approved body credentials where supported.
    workerToken:
      type: apiKey
      in: header
      name: X-Worker-Token
      description: Rotatable internal credential accepted only from explicitly trusted network peers.
  parameters:
...
    UserId: {name: userId, in: path, required: true, schema: {type: string, format: uuid}}
    ClientId: {name: clientId, in: path, required: true, schema: {type: string}}
...
    ProviderId: {name: providerId, in: path, required: true, schema: {type: string, format: uuid}}
...
  requestBodies:
...
    MfaOptionalVerification: {required: false, content: {application/json: {schema: {$ref: '#/components/schemas/MfaOptionalVerificationRequest'}}}}
    MfaOptionalVerificationRequired: {required: true, content: {application/json: {schema: {$ref: '#/components/schemas/MfaOptionalVerificationRequest'}}}}
...
    AdminUpdateRole: {required: true, content: {application/json: {schema: {$ref: '#/components/schemas/AdminUpdateRoleRequest'}}}}
    AdminAccountState: {required: true, content: {application/json: {schema: {$ref: '#/components/schemas/AdminAccountStateRequest'}}}}
    AdminOAuthClientCreate: {required: true, content: {application/json: {schema: {$ref: '#/components/schemas/AdminOAuthClientCreateRequest'}}}}
    AdminOAuthClientUpdate: {required: true, content: {application/json: {schema: {$ref: '#/components/schemas/AdminOAuthClientUpdateRequest'}}}}
    AdminSocialProviderUpdate: {required: true, content: {application/json: {schema: {$ref: '#/components/schemas/AdminSocialProviderUpdateRequest'}}}}
  headers:
    XRequestId:
      required: true
      schema: {type: string, format: uuid}
      description: Request correlation identifier; a valid inbound X-Request-ID is preserved, otherwise AuthKit generates one.
  responses:
    Success:
      description: Successful response
      headers: {X-Request-ID: {$ref: '#/components/headers/XRequestId'}}
      content: {application/json: {schema: {$ref: '#/components/schemas/ApiResponse'}}}
    Error:
      description: Opaque error response
      headers: {X-Request-ID: {$ref: '#/components/headers/XRequestId'}}
      content: {application/json: {schema: {$ref: '#/components/schemas/ApiResponse'}}}
...
  schemas:
...
    AdminUpdateRoleRequest:
      type: object
      additionalProperties: false
      required: [role]
      properties:
        role: {type: string, enum: [USER, PLATFORM_ADMIN]}
        currentPassword: {type: string, maxLength: 128, nullable: true, writeOnly: true}
        mfaCode: {type: string, maxLength: 32, nullable: true, writeOnly: true}
    AdminAccountStateRequest:
      type: object
      additionalProperties: false
      required: [reason]
      properties:
        reason: {type: string, minLength: 1, maxLength: 500}
        currentPassword: {type: string, maxLength: 128, nullable: true, writeOnly: true}
        mfaCode: {type: string, maxLength: 32, nullable: true, writeOnly: true}
    AdminOAuthClientCreateRequest:
      type: object
      additionalProperties: false
      required: [displayName, publicClient, redirectUris, scopes, requirePkce]
      properties:
        displayName: {type: string, minLength: 1, maxLength: 120}
        publicClient: {type: boolean}
        redirectUris: {type: array, minItems: 1, maxItems: 20, uniqueItems: true, items: {type: string, format: uri, maxLength: 512}}
        scopes: {type: array, minItems: 1, maxItems: 20, uniqueItems: true, items: {type: string, maxLength: 80}}
        requirePkce: {type: boolean}
        currentPassword: {type: string, maxLength: 128, nullable: true, writeOnly: true}
        mfaCode: {type: string, maxLength: 32, nullable: true, writeOnly: true}
    AdminOAuthClientUpdateRequest:
      type: object
      additionalProperties: false
      required: [displayName, redirectUris, scopes, requirePkce, rotateSecret]
      properties:
        displayName: {type: string, minLength: 1, maxLength: 120}
        redirectUris: {type: array, minItems: 1, maxItems: 20, uniqueItems: true, items: {type: string, format: uri, maxLength: 512}}
        scopes: {type: array, minItems: 1, maxItems: 20, uniqueItems: true, items: {type: string, maxLength: 80}}
        requirePkce: {type: boolean}
        rotateSecret: {type: boolean, description: Returns a new secret exactly once for confidential clients}
        currentPassword: {type: string, maxLength: 128, nullable: true, writeOnly: true}
        mfaCode: {type: string, maxLength: 32, nullable: true, writeOnly: true}
    AdminSocialProviderUpdateRequest:
      type: object
      additionalProperties: false
      required: [displayName, clientId, scopes, clientAuthMethod]
      properties:
        displayName: {type: string, minLength: 1, maxLength: 120}
        clientId: {type: string, minLength: 1, maxLength: 255}
        clientSecret: {type: string, maxLength: 2048, nullable: true, writeOnly: true}
        scopes: {type: array, minItems: 1, maxItems: 20, uniqueItems: true, items: {type: string, maxLength: 80}}
        clientAuthMethod: {type: string, enum: [CLIENT_SECRET_BASIC, CLIENT_SECRET_POST]}
        currentPassword: {type: string, maxLength: 128, nullable: true, writeOnly: true}
        mfaCode: {type: string, maxLength: 32, nullable: true, writeOnly: true}
...
    ApiResponse:
      type: object
      additionalProperties: false
      required: [timestamp, requestId]
      properties:
        data: {}
        errors: {type: array, items: {type: string}, nullable: true}
        timestamp: {type: string, format: date-time}
        code: {type: string, pattern: '^[a-z0-9_]+$', nullable: true, description: Stable machine-readable error code}
        requestId: {type: string, minLength: 1, maxLength: 128}
...
    MfaOptionalVerificationRequest:
      type: object
      additionalProperties: false
      properties:
        currentPassword: {type: string, maxLength: 128, nullable: true, writeOnly: true}
        code: {type: string, minLength: 6, maxLength: 32, nullable: true, writeOnly: true}
    AdminSocialProviderCreateRequest:
      type: object
      additionalProperties: false
      required: [providerKey, displayName, providerType, issuer, clientId, clientSecret, scopes, clientAuthMethod]
      properties:
        providerKey: {type: string, pattern: '^[a-z0-9][a-z0-9_-]{1,63}$'}
        displayName: {type: string, maxLength: 120}
        providerType: {type: string, enum: [GOOGLE, GENERIC_OIDC, GOVBR_EXPERIMENTAL]}
        issuer: {type: string, format: uri, description: Exact HTTPS issuer from the operator allowlist}
        clientId: {type: string, maxLength: 255}
        clientSecret: {type: string, maxLength: 2048, writeOnly: true}
        scopes: {type: array, minItems: 1, maxItems: 20, items: {type: string}}
        clientAuthMethod: {type: string, enum: [CLIENT_SECRET_BASIC, CLIENT_SECRET_POST]}
        currentPassword: {type: string, maxLength: 128, writeOnly: true}
        mfaCode: {type: string, maxLength: 32, writeOnly: true}
...